services/main-bot/requirements.txt
python-telegram-bot==20.7
httpx==0.26.0
aiokafka==0.10.0
